# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Chart API - Alpaca market data proxy
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Bar {
  /** Timestamp (ISO 8601) */
  t: string;
  /** Open price */
  o: number;
  /** High price */
  h: number;
  /** Low price */
  l: number;
  /** Close price */
  c: number;
  /** Volume */
  v: number;
  /** Volume weighted average price */
  vw?: number;
}

export interface BarsResponse {
  symbol: string;
  bars: Bar[];
  nextPageToken?: string | null;
}

export interface SymbolInfo {
  symbol: string;
  name: string;
  exchange: string;
  type: string;
}

export interface SearchResponse {
  results: SymbolInfo[];
}

export interface QuoteResponse {
  symbol: string;
  price: number;
  change: number;
  changePercent: number;
  open?: number;
  high?: number;
  low?: number;
  volume?: number;
  timestamp: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetBarsParams = {
  /**
   * Stock or crypto ticker symbol (e.g. AAPL, BTCUSD)
   */
  symbol: string;
  /**
   * Bar timeframe
   */
  timeframe: GetBarsTimeframe;
  /**
   * Start datetime (ISO 8601)
   */
  start?: string;
  /**
   * End datetime (ISO 8601)
   */
  end?: string;
  /**
   * Number of bars to return
   */
  limit?: number;
  /**
   * Data feed (sip requires premium)
   */
  feed?: GetBarsFeed;
};

export type GetBarsTimeframe =
  (typeof GetBarsTimeframe)[keyof typeof GetBarsTimeframe];

export const GetBarsTimeframe = {
  "1Min": "1Min",
  "5Min": "5Min",
  "15Min": "15Min",
  "30Min": "30Min",
  "1Hour": "1Hour",
  "4Hour": "4Hour",
  "1Day": "1Day",
  "1Week": "1Week",
} as const;

export type GetBarsFeed = (typeof GetBarsFeed)[keyof typeof GetBarsFeed];

export const GetBarsFeed = {
  iex: "iex",
  sip: "sip",
  otc: "otc",
} as const;

export type SearchSymbolsParams = {
  /**
   * Search query string
   */
  query: string;
};

export type GetQuoteParams = {
  symbol: string;
};