# lib/api-zod/src/generated/types/barsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Chart API - Alpaca market data proxy
 * OpenAPI spec version: 0.1.0
 */
import type { Bar } from "./bar";

export interface BarsResponse {
  symbol: string;
  bars: Bar[];
  nextPageToken?: string | null;
}