# lib/api-zod/src/generated/types/quoteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Chart API - Alpaca market data proxy
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteResponse {
  symbol: string;
  price: number;
  change: number;
  changePercent: number;
  open?: number;
  high?: number;
  low?: number;
  volume?: number;
  timestamp: string;
}