# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Chart API - Alpaca market data proxy
 * OpenAPI spec version: 0.1.0
 */
import type { SymbolInfo } from "./symbolInfo";

export interface SearchResponse {
  results: SymbolInfo[];
}