# lib/api-zod/src/generated/types/symbolInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trading Chart API - Alpaca market data proxy
 * OpenAPI spec version: 0.1.0
 */

export interface SymbolInfo {
  symbol: string;
  name: string;
  exchange: string;
  type: string;
}